Pass IP network type from network to DevicePortConfig
Description:
Network configuration that comes from cloud already has a network type
field (ipv4/ipv6). This field is currently not being passed from
zedagent to NIM as part of the NetworkPortConfig in DevicePortConfig.

This commit pushes the network type field from zedagent to NIM as part
of NetworkPortConfig. We use the IP type to run DHCP for specific
protocol type.

EVE today runs in Dual stack mode irrespective of the IP family of
network. There may be users benefitting from the current behavior.
It will good not to surprise the current users. Idea is to keep the
current dual stack behavior and instroduce new types V4Only & V6Only
that can be used if users want the uplink adapters to only have V4 or
only V6 address families.

Signed-off-by: Gopi krishna Kodali <gkodali@zededa.com>